templates/fastapi-ml/requirements.txt
fastapi
uvicorn[standard]
scikit-learn
pandas
numpy
pydantic
python-dotenv
jupyter
